Add an optional console_script entry point which gets enabled by the `keyring` extra.
Looking for a `keyring` executable on the PATH proved to be problematic. Poetry has keyring as a dependency. Resulting in it finding Poetry's keyring executable before the one I want it to find.

Now our own `keyring-subprocess` console_script entry point, which wraps keyring's console_script entry point, is the landmark file that will be queried when found.